add IntModel and reorganize models code files
this also resulted in a change in the API. users now initialize the
model and then call model.FromPlainFile() to load a plaintext file.
type-specific functionality is in methods, while 'Model' interface
generic functionality are in functions which take model as an argument.
go does not allow interfaces to be receivers.